+++ docs/changelog.rst
@@ -90,6 +90,8 @@ Little improvements and fixes:
   work in ``auto`` mode. Thanks to Harry Khanna.
 * The :ref:`write-cmd` command now has a ``--force`` flag. Thanks again to
   Harry Khanna.
+* :doc:`/plugins/mbsync`: Track alignment now works with albums that have
+  multiple copies of the same recording. Thanks to Rui Gonçalves.
 
 
 1.3.6 (May 10, 2014)